Switched to synchronous management of onboarding dialog (#52)
Replaced OnboardingDialogClippedBackground with OnboardingDialogWithHighlight, a synchronous version of the same widget : FutureBuilder logic is now managed by a ValueListenableBuilder and synchronicity with navigation events no longer relies on navigation listeners. Instead, an internal stability check makes the decision to update the drawn rect. This change allows the OnboardingDialog to draw its rect regardless of the navigation event occuring (page transition, drawer opening, etc.) It now tracks internally the stability of the rect to be drawn and  displays it once it is stable.

Added button interaction blockers (_navEnabled boolean)  to prevent the user to skip several onboarding steps at once.

OnboardingDialogClippedBackground remains accessible as an asynchronous onboardingDialog management option. Added maxAttempts int to prevent loop access to undisposed future when trying to refresh futureBuilder.